chore(components-stories): re-organized the components looks on the documentation page
